+++ Vi/Podfile
@@ -5,16 +5,19 @@ target 'Vi' do
   pod 'FontAwesome.swift'
   pod 'AFNetworking', '~> 2.5'
   pod 'Lock', '~> 1.2'
+  pod 'JWTDecode', '~> 1.0'
 end
 
 target 'ViTests' do
   pod 'FontAwesome.swift'
   pod 'AFNetworking', '~> 2.5'
   pod 'Lock', '~> 1.2'
+  pod 'JWTDecode', '~> 1.0'
 end
 
 target 'ViUITests' do
   pod 'FontAwesome.swift'
   pod 'AFNetworking', '~> 2.5'
   pod 'Lock', '~> 1.2'
+  pod 'JWTDecode', '~> 1.0'
 end
