+++ Vi/Vi/Base.lproj/Library.storyboard
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="8191" systemVersion="15A284" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="9059" systemVersion="15A284" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
     <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="8154"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9049"/>
         <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
     </dependencies>
     <customFonts key="customFonts">
@@ -173,7 +173,11 @@
                         </variation>
                     </view>
                     <navigationItem key="navigationItem" id="WQz-Ti-LcY">
-                        <barButtonItem key="rightBarButtonItem" title="Sign In" id="9do-UI-6uy"/>
+                        <barButtonItem key="rightBarButtonItem" title="Sign In" id="9do-UI-6uy">
+                            <connections>
+                                <action selector="signIn:" destination="4Zl-Hm-WZ5" id="jvW-SA-gEN"/>
+                            </connections>
+                        </barButtonItem>
                     </navigationItem>
                     <connections>
                         <outlet property="CommandView" destination="goS-Rs-0lT" id="HCR-h9-F50"/>
